Refactoring finished
