+++ setup.py
@@ -1,0 +1,13 @@
+from setuptools import setup, find_packages
+
+
+setup(
+    name='unires',
+    version='0.1a',
+    packages=find_packages(),
+    url='https://github.com/brudfors/UniRes',
+    author='Mikael Brudfors',
+    author_email='brudfors@gmail.com',
+    description='UniRes: Unified Super-Resolution of Neuroimaging Data',
+    python_requires='>=3.5',
+)
