+++ unires.egg-info/PKG-INFO
@@ -1,0 +1,11 @@
+Metadata-Version: 1.2
+Name: unires
+Version: 0.1a0
+Summary: UniRes: Unified Super-Resolution of Neuroimaging Data
+Home-page: https://github.com/brudfors/UniRes
+Author: Mikael Brudfors
+Author-email: brudfors@gmail.com
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
+Requires-Python: >=3.5
